Add Data::from/to_xml_format
This mirrors the similar methods on the `Date` type.

I'm in a funny situation where I need to construct a plist that is
embedded in another XML document; I'm doing this by writing custom
derive impls for the various Value types, and having these methods live
here saves me from needing to manage the base64 dependency myself.

No sweat if this feels overly specific, I have other options this is
just the simplest for me.