Fail deserialisation if input is not completely consumed
The direct constructors for `Value` already perform this check.

I believe this should not cause problems with existing code as the only
XML and ASCII plists could be malformed in this way, and it would be
unlikely to find an XML plist with two children inside the root
`<plist>` element.